refactor location of item descriptor to serialization folder
