--- libs/internal/src/data_model/sdk_data_set.cpp
@@ -1,3 +1,0 @@
-#include <launchdarkly/data_model/sdk_data_set.hpp>
-
-namespace launchdarkly::data_model {}
